pyproject.toml
[tool.poetry]
name = "lda-tg-bot"
version = "0.0.1"
description = "Modular telegram bot with fun Ml tools."
authors = ["GeorgiEstonia <georgi.estonia@gmail.com>"]
readme = "README.md"
packages = [{include = "lda_tg_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
replicate = "^0.4.0"
python-decouple = "^3.6"
numpy = "^1.23.4"
nltk = "^3.7"
gensim = "^4.2.0"
pandas = "^1.5.1"
scrapfly-sdk = "^0.8.2"
flask = "^2.2.2"
flask-cors = "^3.0.10"
six = "^1.16.0"
gunicorn = "^20.1.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
mypy = "^0.982"
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
